refactor: remove UK (GB) country code support and related components
