--- src/app/gb/layout.tsx
@@ -1,69 +1,0 @@
-import { SpeedInsights } from '@vercel/speed-insights/next'
-import { capitalize } from 'lodash-es'
-import type { Metadata } from 'next'
-import NextTopLoader from 'nextjs-toploader'
-
-import { TopLevelClientLogic } from '@/app/topLevelClientLogic'
-import { Footer } from '@/components/app/footer'
-import { GoogleTagManager } from '@/components/app/googleTagManager'
-import { Navbar } from '@/components/app/navbar'
-import { NavBarGlobalBanner } from '@/components/app/navbarGlobalBanner'
-import { OverrideGlobalLocalStorage } from '@/components/app/overrideGlobalLocalStorage'
-import { FullHeight } from '@/components/ui/fullHeight'
-import { Toaster } from '@/components/ui/sonner'
-import { getOpenGraphImageUrl } from '@/utils/server/generateOpenGraphImageUrl'
-import { generateMetadataDetails, TOP_LEVEL_METADATA_DETAILS } from '@/utils/server/metadataUtils'
-import { NEXT_PUBLIC_ENVIRONMENT } from '@/utils/shared/sharedEnv'
-import { COUNTRY_CODE_TO_LOCALE, SupportedCountryCodes } from '@/utils/shared/supportedCountries'
-import { fontClassName } from '@/utils/web/fonts'
-
-export { viewport } from '@/utils/server/metadataUtils'
-
-export async function generateStaticParams() {
-  return SupportedCountryCodes.GB
-}
-
-const title = `${
-  NEXT_PUBLIC_ENVIRONMENT === 'production'
-    ? ''
-    : `${capitalize(NEXT_PUBLIC_ENVIRONMENT.toLowerCase())} Env - `
-}Stand With Crypto`
-const description = `Stand With Crypto Alliance is a non-profit organization dedicated to uniting global crypto advocates.`
-const ogImage = getOpenGraphImageUrl({ title: description })
-
-export const metadata: Metadata = {
-  ...generateMetadataDetails({ description, title, ogImage }),
-  title: {
-    default: title,
-    template: '%s | Stand With Crypto',
-  },
-  ...TOP_LEVEL_METADATA_DETAILS,
-}
-
-export default async function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang={COUNTRY_CODE_TO_LOCALE[SupportedCountryCodes.GB]} translate="no">
-      <GoogleTagManager />
-      <body className={fontClassName}>
-        <OverrideGlobalLocalStorage />
-        <NextTopLoader
-          color="hsl(var(--primary-cta))"
-          shadow="0 0 10px hsl(var(--primary-cta)),0 0 5px hsl(var(--primary-cta))"
-          showSpinner={false}
-        />
-        <TopLevelClientLogic countryCode={SupportedCountryCodes.GB}>
-          <FullHeight.Container>
-            <NavBarGlobalBanner />
-            <Navbar countryCode={SupportedCountryCodes.GB} />
-            <FullHeight.Content>{children}</FullHeight.Content>
-            <Footer countryCode={SupportedCountryCodes.GB} />
-          </FullHeight.Container>
-        </TopLevelClientLogic>
-        <Toaster />
-        {/* TODO(@twistershark): Create specific CookieConsent for GB */}
-        {/* <CookieConsent countryCode={SupportedCountryCodes.GB} /> */}
-        <SpeedInsights debug={false} sampleRate={0.04} />
-      </body>
-    </html>
-  )
-}
