+++ src/utils/shared/supportedCountries.ts
@@ -2,19 +2,15 @@ import { SupportedLocale } from '@/utils/shared/supportedLocales'
 
 export enum SupportedCountryCodes {
   US = 'us',
-  GB = 'gb',
 }
 
 export const DEFAULT_SUPPORTED_COUNTRY_CODE = SupportedCountryCodes.US
 export const ORDERED_SUPPORTED_COUNTRIES: readonly SupportedCountryCodes[] = [
   SupportedCountryCodes.US,
-  // TODO: uncomment this line when we complete the UK implementation
-  // SupportedCountryCodes.GB,
 ]
 
 export const COUNTRY_CODE_TO_LOCALE: Record<SupportedCountryCodes, SupportedLocale> = {
   [SupportedCountryCodes.US]: SupportedLocale.EN_US,
-  [SupportedCountryCodes.GB]: SupportedLocale.EN_GB,
 }
 
 // Two lowercase letters (e.g., "us", "uk")
